+++ src/api/annotations.ts
@@ -1,8 +1,8 @@
 import { lastValueFrom } from 'rxjs';
 import { Annotation } from 'types/annotation';
-import { FieldType, MutableDataFrame } from '@grafana/data';
-import { getBackendSrv } from '@grafana/runtime';
-import { Messages, RequestTypeValue } from '../constants';
+import { FieldType, MutableDataFrame, TimeRange } from '@grafana/data';
+import { getBackendSrv, getTemplateSrv } from '@grafana/runtime';
+import { AnnotationDashboard, AnnotationRange, AnnotationType, Messages, RequestType } from '../constants';
 import { Query } from '../types';
 import { notifyError } from '../utils';
 import { Api } from './api';
@@ -10,10 +10,52 @@ import { Api } from './api';
 /**
  * Get Annotations
  */
-export const getAnnotations = async (api: Api): Promise<Annotation[]> => {
+export const getAnnotations = async (
+  api: Api,
+  query: Query,
+  range: TimeRange,
+  dashboardUID: string | undefined
+): Promise<Annotation[]> => {
+  let params: Record<string, any> = {};
+
+  /**
+   * Time Range
+   */
+  if (query.annotationRange === AnnotationRange.SELECTED) {
+    params.from = range.from.valueOf();
+    params.to = range.to.valueOf();
+  }
+
+  /**
+   * Dashboard
+   */
+  if (query.annotationDashboard === AnnotationDashboard.THIS && dashboardUID) {
+    params.dashboardUID = dashboardUID;
+  }
+
+  /**
+   * Max Limit
+   */
+  if (query.annotationLimit) {
+    params.limit = query.annotationLimit;
+  }
+
+  /**
+   * Filter Type
+   */
+  if (query.annotationType === AnnotationType.ALERT) {
+    params.type = AnnotationType.ALERT;
+  } else if (query.annotationType === AnnotationType.ANNOTATION) {
+    params.type = AnnotationType.ANNOTATION;
+  }
+
+  /**
+   * Fetch
+   */
   const response = await lastValueFrom(
     getBackendSrv().fetch({
       method: 'GET',
+      params,
       url: `${api.instanceSettings.url}/api/annotations`,
     })
   );
@@ -27,14 +69,29 @@ export const getAnnotations = async (api: Api): Promise<Annotation[]> => {
     return [];
   }
 
-  return response.data as Annotation[];
+  let annotations = response.data as Annotation[];
+
+  /**
+   * Filter Pattern
+   */
+  if (query.annotationPattern) {
+    const pattern = getTemplateSrv().replace(query.annotationPattern);
+    annotations = annotations.filter((annotation) => annotation.text?.match(pattern));
+  }
+
+  return annotations;
 };
 
 /**
  * Get Annotations Frame
  */
-export const getAnnotationsFrame = async (api: Api, query: Query): Promise<MutableDataFrame[]> => {
-  const annotations = await getAnnotations(api);
+export const getAnnotationsFrame = async (
+  api: Api,
+  query: Query,
+  range: TimeRange,
+  dashboardUID: string | undefined
+): Promise<MutableDataFrame[]> => {
+  const annotations = await getAnnotations(api, query, range, dashboardUID);
   if (!annotations.length) {
     return [];
   }
@@ -43,7 +100,7 @@ export const getAnnotationsFrame = async (api: Api, query: Query): Promise<Mutab
    * Create frame
    */
   const frame = new MutableDataFrame({
-    name: RequestTypeValue.ANNOTATIONS,
+    name: RequestType.ANNOTATIONS,
     refId: query.refId,
     fields: [
       {