+++ src/types/query.ts
@@ -1,5 +1,5 @@
 import { DataQuery } from '@grafana/data';
-import { RequestTypeValue } from '../constants';
+import { AnnotationDashboard, AnnotationRange, AnnotationType, RequestType } from '../constants';
 
 /**
  * Query
@@ -8,7 +8,42 @@ export interface Query extends DataQuery {
   /**
    * Request Type
    *
-   * @type {RequestTypeValue}
+   * @type {RequestType}
    */
-  requestType?: RequestTypeValue;
+  requestType?: RequestType;
+
+  /**
+   * Annotation Type
+   *
+   * @type {AnnotationType}
+   */
+  annotationType?: AnnotationType;
+
+  /**
+   * Annotation Dashboard
+   *
+   * @type {AnnotationDashboard}
+   */
+  annotationDashboard?: AnnotationDashboard;
+
+  /**
+   * Annotation Range
+   *
+   * @type {AnnotationRange}
+   */
+  annotationRange?: AnnotationRange;
+
+  /**
+   * Annotation Pattern
+   *
+   * @type {string}
+   */
+  annotationPattern?: string;
+
+  /**
+   * Annotation Limit
+   *
+   * @type {number}
+   */
+  annotationLimit?: number;
 }